Added language support to chat API

+++ chatbot/src/main/java/com/example/chatbot/entity/Player.java
@@ -12,7 +12,7 @@ public class Player {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String firstName;
 	private String lastName;
 
@@ -21,13 +21,19 @@ public class Player {
 		this.lastName = lastName;
 	}
 
-	public Integer getId() {
+	
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+
+
+	public void setId(Long id) {
 		this.id = id;
 	}
+
+
 
 	public String getFirstName() {
 		return firstName;
